linknode: move store from channeldb
Move link-node persistence into its own package while keeping channeldb
aliases and wrappers for existing callers.

This creates an explicit linknode.Store boundary that can be wired
independently from channel-state persistence in later commits.
